+++ src/analysis/invglot/iaif.cpp
@@ -31,7 +31,7 @@ static std::vector<double> calculateLPC(const std::vector<double>& x, const std:
         lpcIn[i] = w[i] * x[x.size() / 2 - len / 2 + i];
     }
     auto a = lpc->solve(lpcIn.data(), len, order, &gain);
-    a.insert(a.begin(), 1.0f);
+    a.insert(a.begin(), 1.0);
     return a;
 }
 
@@ -43,7 +43,7 @@ static std::vector<double> removePreRamp(const std::vector<double>& x, int prefl
 inline double G(double x, int L, double alpha)
 {
     const int N = L - 1;
-    const double k = (x - N / 2.0f) / (2 * L * alpha);
+    const double k = (x - N / 2.0) / (2 * L * alpha);
     return expf(-(k * k));
 }
 
@@ -51,8 +51,8 @@ static void calcGaussian(std::vector<double>& win, double alpha)
 {
     const int L = win.size();
     
-    double Gmh = G(-0.5f, L, alpha);
-    double GmhpLpGmhmL = G(-0.5f + L, L, alpha) - G(-0.5f - L, L, alpha);
+    double Gmh = G(-0.5, L, alpha);
+    double GmhpLpGmhmL = G(-0.5 + L, L, alpha) - G(-0.5 - L, L, alpha);
 
     for (int n = 0; n < L; ++n) {
         win[n] = G(n, L, alpha) - (Gmh * (G(n + L, L, alpha) + G(n - L, L, alpha))) / GmhpLpGmhmL;
@@ -64,16 +64,16 @@ InvglotResult IAIF::solve(const double *xData, int length, double sampleRate)
     const int p_gl = 2;
     const int p_vt = 2 * std::round(sampleRate / 2000) + 4;
 
-    const int lpW = std::round(0.015f * sampleRate);
+    const int lpW = std::round(0.015 * sampleRate);
 
-    std::vector<double> one({1.0f});
-    std::vector<double> oneMinusD({1.0f, -d});
+    std::vector<double> one({1.0});
+    std::vector<double> oneMinusD({1.0, -d});
 
     static std::vector<double> window;
     if (window.size() != lpW) {
         window.resize(lpW);
         for (int i = 0; i < lpW; ++i) {
-            window[i] = 0.5f - 0.5f * cosf((2.0f * M_PI * i) / (double) (length - 1));
+            window[i] = 0.5 - 0.5 * cos((2.0 * M_PI * i) / (double) (length - 1));
         }
     }
 
@@ -81,14 +81,14 @@ InvglotResult IAIF::solve(const double *xData, int length, double sampleRate)
    
     static std::vector<std::array<double, 6>> hpfilt;
     if (hpfilt.empty()) {
-        hpfilt = Analysis::butterworthHighpass(8, 70.0f, sampleRate);
+        hpfilt = Analysis::butterworthHighpass(8, 70.0, sampleRate);
     }
 
     int preflt = p_vt + 1;
 
     std::vector<double> xWithPreRamp(preflt + length);
     for (int i = 0; i < preflt; ++i) {
-        xWithPreRamp[i] = 2.0f * ((double) i / (double) (preflt - 1) - 0.5f) * x[0];
+        xWithPreRamp[i] = 2.0 * ((double) i / (double) (preflt - 1) - 0.5) * x[0];
     }
     for (int i = 0; i < length; ++i) {
         xWithPreRamp[preflt + i] = x[i];
